Add ability to set 'current_default' for model
Previously when accessing a record either as an automatic parent or
through `#the_model_name` method it was always the first created record
of the given type that was returned.

Now you can set the `current_default` for a given model. This is useful
if you want to associate a number of records with one parent and then
another bunch of records to another parent of the same type.